requirements.txt
lxml==4.6.4
pyzmq==22.3.0
token-bucket==0.3.0
pytest==6.2.5
youtube-search=2.1.0
